Moved tests related to Surefire-api dependency to separated class
